pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image_denoiser"
version = "0.0.1"
authors = [
  { name="Georges Labrèche", email="georges@tanagraspace.com" },
  { name="Dr. Cesar Guzman", email="cesar.guzman@tanagraspace.com" },
]
description = "Denoising OPS-SAT images with TensorFlow Autoencoders."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    'tensorflow==2.12.0', 'matplotlib==3.7.1', 'pandas==2.0.1', 'Pillow==9.5.0'
]

[project.urls]
"Homepage" = "https://github.com/georgeslabreche/opssat-image-denoiser"
"Bug Tracker" = "https://github.com/georgeslabreche/opssat-image-denoiser/issues"

[project.optional-dependencies]
test = [
    'pytest==6.2.5',
]
